requirements.txt
apache-airflow==2.7.1
pandas
requests
pyarrow